refactor: 펙토리 디렉토리 이름 변경 및 구조 정리
