+++ src/main/java/lotto/domain/lotto/factory/LottoFactory.java
@@ -1,4 +1,4 @@
-package lotto.domain.lottofactory;
+package lotto.domain.lotto.factory;
 
 import lotto.domain.lotto.Lotto;
